switched it from unit8_t to int to allow larger values

+++ src/main.c
@@ -54,7 +54,7 @@
 uint8_t num_repeats = '5';
 char id[] = "BALL1";
 char location_string[] = "52.0937,-1.0161";
-uint8_t tx_gap = 1000; // milliseconds between tx = tx_gap * 100, therefore 1000 = 100seconds
+int tx_gap = 1000; // milliseconds between tx = tx_gap * 100, therefore 1000 = 100seconds
 
 char data_temp[64];
 uint8_t data_count = 95; // 'a' - 1 (as the first function will at 1 to make it 'a'
